[WIP] Remove std requirement from rustzx-core
