+++ rustzx-core/src/lib.rs
@@ -6,5 +6,10 @@ pub mod settings;
 pub mod utils;
 pub mod z80;
 pub mod zx;
+pub mod host;
+pub mod error;
 
 extern crate alloc;
+
+pub use crate::error::Error;
+pub type Result<T> = core::result::Result<T, Error>;
